// src/lib/types/renovation.d.ts
export interface RenovationData {

    applicantName:string;
    block:string;
    detailsWork:string;
    startDate:string;
    endDate:string;
    contractorName:string;
    status:string;
    contractorNumber:string;
    renovationStatus:string;
    personName:string;
    personNumber:string;
    userId:number

}